Localize Qt integration
