--- default-impl/qt-udp-broadcast/CMakeLists.txt
@@ -1,72 +1,0 @@
-#   Copyright (C) 2015-2020 Virgil Security Inc.
-#
-#   All rights reserved.
-#
-#   Redistribution and use in source and binary forms, with or without
-#   modification, are permitted provided that the following conditions are
-#   met:
-#
-#       (1) Redistributions of source code must retain the above copyright
-#       notice, this list of conditions and the following disclaimer.
-#
-#       (2) Redistributions in binary form must reproduce the above copyright
-#       notice, this list of conditions and the following disclaimer in
-#       the documentation and/or other materials provided with the
-#       distribution.
-#
-#       (3) Neither the name of the copyright holder nor the names of its
-#       contributors may be used to endorse or promote products derived from
-#       this software without specific prior written permission.
-#
-#   THIS SOFTWARE IS PROVIDED BY THE AUTHOR ''AS IS'' AND ANY EXPRESS OR
-#   IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-#   WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-#   DISCLAIMED. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT,
-#   INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-#   (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-#   SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
-#   HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
-#   STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING
-#   IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-#   POSSIBILITY OF SUCH DAMAGE.
-#
-#   Lead Maintainer: Virgil Security Inc. <support@virgilsecurity.com>
-
-cmake_minimum_required(VERSION 3.11)
-
-set(CMAKE_AUTOMOC ON)
-set(CMAKE_AUTORCC ON)
-
-find_package(Qt5 COMPONENTS Core REQUIRED)
-find_package(Qt5 COMPONENTS Network REQUIRED)
-
-add_library(iotkit-qt-netif-udp-broadcast)
-
-target_sources(iotkit-qt-netif-udp-broadcast
-        PRIVATE
-
-        # Headers
-        ${CMAKE_CURRENT_LIST_DIR}/include/virgil/iot/qt-udp-broadcast/VSQUdpBroadcast.h
-
-        # Sources
-        ${CMAKE_CURRENT_LIST_DIR}/src/VSQUdpBroadcast.cpp
-        )
-
-target_include_directories(iotkit-qt-netif-udp-broadcast
-        PUBLIC
-        ${CMAKE_CURRENT_LIST_DIR}/include
-        )
-
-target_link_libraries(iotkit-qt-netif-udp-broadcast
-        PUBLIC
-
-        iotkit-qt
-        vs-module-snap-control
-
-        Qt5::Core
-        Qt5::Network
-        )
-
-if(COMMAND add_clangformat)
-    add_clangformat(iotkit-qt-netif-udp-broadcast)
-endif()
