+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQDeviceRoles.h
