+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQDeviceSerial.h
