+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQDeviceType.h
