+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQFeatures.h
