+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQFileVersion.h
