+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQImplementations.h
