+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQIoTKitFacade.h
@@ -1,0 +1,75 @@
+//  Copyright (C) 2015-2020 Virgil Security, Inc.
+//
+//  All rights reserved.
+//
+//  Redistribution and use in source and binary forms, with or without
+//  modification, are permitted provided that the following conditions are
+//  met:
+//
+//      (1) Redistributions of source code must retain the above copyright
+//      notice, this list of conditions and the following disclaimer.
+//
+//      (2) Redistributions in binary form must reproduce the above copyright
+//      notice, this list of conditions and the following disclaimer in
+//      the documentation and/or other materials provided with the
+//      distribution.
+//
+//      (3) Neither the name of the copyright holder nor the names of its
+//      contributors may be used to endorse or promote products derived from
+//      this software without specific prior written permission.
+//
+//  THIS SOFTWARE IS PROVIDED BY THE AUTHOR ''AS IS'' AND ANY EXPRESS OR
+//  IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+//  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+//  DISCLAIMED. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT,
+//  INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+//  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+//  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+//  HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+//  STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING
+//  IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+//  POSSIBILITY OF SUCH DAMAGE.
+//
+//  Lead Maintainer: Virgil Security Inc. <support@virgilsecurity.com>
+
+#ifndef VSQIOTKITFACADE_H
+#define VSQIOTKITFACADE_H
+
+#include <QtCore>
+#include <QtNetwork>
+#include <virgil/iot/logger/logger.h>
+#include <virgil/iot/provision/provision-structs.h>
+#include <virgil/iot/qt/helpers/VSQSingleton.h>
+#include <virgil/iot/qt/helpers/VSQFeatures.h>
+#include <virgil/iot/qt/helpers/VSQImplementations.h>
+#include <virgil/iot/qt/helpers/VSQDeviceRoles.h>
+#include <virgil/iot/qt/helpers/VSQAppConfig.h>
+
+class VSQSnapInfoClient;
+class VSQSnapServiceBase;
+
+class VSQIoTKitFacade : public QObject, public VSQSingleton<VSQIoTKitFacade> {
+    Q_OBJECT
+
+public:
+    bool
+    init(const VSQFeatures &features, const VSQImplementations &impl, const VSQAppConfig &appConfig);
+
+private slots:
+    void
+    restartInfoClientPolling(QAbstractSocket::SocketState connectionState);
+
+private:
+    VSQFeatures m_features;
+    VSQImplementations m_impl;
+    VSQAppConfig m_appConfig;
+    QSharedPointer<VSQSnapInfoClient> m_serviceInfoClient;
+
+    void
+    initSnap();
+
+    void
+    registerService(VSQSnapServiceBase &service);
+};
+
+#endif // VSQIOTKITFACADE_H
