+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQMac.h
