+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQManufactureId.h
