+++ integration/qt/facade/include/virgil/iot/qt/helpers/VSQSingleton.h
